fix: Fixed age field not being set by jpa
Jpa uses the non-args constructor to create the entity, that's why it was not computing the age field value (not included in db), but was setting the others  field values just fine (stored in db).
The @PostLoad annotation executes the method after the entity is loaded, and by doing this I can calculate and set the age value.